// src/types/xml/IMeter.ts
export type MeterUnitsOfMeasure =
  | "ccf (hundred cubic feet)"
  | "cf (cubic feet)"
  | "cGal (hundred gallons) (UK)"
  | "cGal (hundred gallons) (US)"
  | "Cubic Meters per Day"
  | "cm (Cubic meters)"
  | "Cords"
  | "Gallons (UK)"
  | "Gallons (US)"
  | "GJ"
  | "kBtu (thousand Btu)"
  | "kcf (thousand cubic feet)"
  | "Kcm (Thousand Cubic meters)"
  | "KGal (thousand gallons) (UK)"
  | "KGal (thousand gallons) (US)"
  | "Kilogram"
  | "KLbs. (thousand pounds)"
  | "kWh (thousand Watt-hours)"
  | "Liters"
  | "MBtu (million Btu)"
  | "MCF(million cubic feet)"
  | "mg/l (milligrams per liter)"
  | "MGal (million gallons) (UK)"
  | "MGal (million gallons) (US)"
  | "Million Gallons per Day"
  | "MLbs. (million pounds)"
  | "MWh (million Watt-hours)"
  | "pounds"
  | "Pounds per year"
  | "therms"
  | "ton hours"
  | "Tonnes (metric)"
  | "tons";

export type TypeOfMeter =
  | "Coal Anthracite"
  | "Coal Bituminous"
  | "Coke"
  | "Diesel"
  | "District Chilled Water - Absorption Chiller using Natural Gas"
  | "District Chilled Water - Electric-Driven Chiller"
  | "District Chilled Water - Engine-Driven Chiller using Natural Gas"
  | "District Chilled Water - Other"
  | "District Hot Water"
  | "District Steam"
  | "Electric"
  | "Electric on Site Solar"
  | "Electric on Site Wind"
  | "Fuel Oil No 1"
  | "Fuel Oil No 2"
  | "Fuel Oil No 4"
  | "Fuel Oil No 5 or 6"
  | "Kerosene"
  | "Natural Gas"
  | "Other (Energy)"
  | "Propane"
  | "Wood"
  | "IT Equipment Input Energy (meters on each piece of equipment)"
  | "Power Distribution Unit (PDU) Input Energy"
  | "Power Distribution Unit (PDU) Output Energy"
  | "Uninterruptible Power Supply (UPS) Output Energy"
  | "Municipally Supplied Potable Water - Mixed Indoor/Outdoor"
  | "Municipally Supplied Potable Water - Indoor"
  | "Municipally Supplied Potable Water - Outdoor"
  | "Municipally Supplied Reclaimed Water - Mixed Indoor/Outdoor"
  | "Municipally Supplied Reclaimed Water - Indoor"
  | "Municipally Supplied Reclaimed Water - Outdoor"
  | "Other - Mixed Indoor/Outdoor (Water)"
  | "Other - Outdoor"
  | "Other - Indoor"
  | "Average Influent Flow"
  | "Well Water - Mixed Indoor/Outdoor"
  | "Well Water - Indoor"
  | "Well Water - Outdoor";

//A service type used for representing a meter, xml: meterType
export interface IMeter {
  id?: number; // The id of the meter.
  type: TypeOfMeter; // The type of meter (i.e. electric, natural gas, PDU, Indoor, etc.).
  name: string; // The name of the meter. 1-100 characters.
  metered?: boolean; // default: true, Indicates if the meter is set up to be metered monthly or for bulk delivery.
  unitOfMeasure: MeterUnitsOfMeasure; // The units that measure the energy for the meter (Kbtu, KWh, Mbtu, MWh, ccf, gallons, etc.).
  firstBillDate: Date; //The date of the first bill.
  inUse: boolean; // Is this meter in use?
  inactiveDate?: Date; // If the meter is no longer in use, this is the date that it went inactive.
  otherDescription?: string; // 1-100 characters. If the type of meter is "Other," this is the description of the meter's energy type.
  accessLevel?: ShareLevel; // Current share level for the user calling the webservice.
  aggregateMeter?: boolean; // If the meter is aggregate meter or not.
  audit?: IAudit;
}